# backend/openapi.yaml
openapi: 3.0.0
info:
  title: QQM API
  version: 1.0.0
  description: Client / project / infrastructure management API. Generated from
    the actual implemented routes, controllers, and zod validators
    (backend/src/openapi) — not from any prior design document.
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: /
    description: Same-origin backend
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT obtained from POST /api/auth/login, sent as `Authorization:
        Bearer <token>`."
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
            details:
              type: object
              additionalProperties: true
          required:
            - message
      required:
        - error
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - INTERNAL_ERROR
    ClientListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - data
        - pagination
    Client:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          nullable: true
          format: date-time
      required:
        - id
        - name
        - status
        - description
        - created_at
        - updated_at
        - deleted_at
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
      required:
        - page
        - per_page
        - total
        - total_pages
...
    PeopleType:
      type: string
      enum:
        - internal_engineer
        - vendor
        - client_contact
        - project_owner
        - approver
...
    PersonListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - data
        - pagination
    Person:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        type:
          $ref: "#/components/schemas/PeopleType"
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          nullable: true
          format: date-time
      required:
        - id
        - name
        - email
        - phone
        - type
        - notes
        - created_at
        - updated_at
        - deleted_at
...
    ActivityLogListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ActivityLog"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required:
        - data
        - pagination
    ActivityLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entity_type:
          $ref: "#/components/schemas/EntityType"
        entity_id:
          type: string
          format: uuid
        action:
          $ref: "#/components/schemas/ActivityAction"
        changed_by:
          type: string
          format: uuid
        old_value:
          type: object
          additionalProperties: true
        new_value:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        changed_by_person:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
          required:
            - id
            - name
      required:
        - id
        - entity_type
        - entity_id
        - action
        - changed_by
        - created_at
        - changed_by_person
    EntityType:
      type: string
      enum:
        - client
        - project
        - environment
        - server
        - credential_reference
        - people
        - resource
        - resource_version
        - schedule
        - user
    ActivityAction:
      type: string
      enum:
        - create
        - update
        - delete
        - restore
...
paths:
...
  /api/clients:
    get:
      tags:
        - Clients
      operationId: listClients
      summary: List clients
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 1
            description: "Default: 1"
          required: false
          description: "Default: 1"
          name: page
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            description: "Default: 20, max: 100"
          required: false
          description: "Default: 20, max: 100"
          name: per_page
          in: query
        - schema:
            type: string
            enum:
              - name
              - status
              - created_at
              - updated_at
            description: "Default: name"
          required: false
          description: "Default: name"
          name: sort
          in: query
        - schema:
            type: string
            enum:
              - asc
              - desc
            description: "Default: asc"
          required: false
          description: "Default: asc"
          name: order
          in: query
        - schema:
            type: string
            enum:
              - "false"
              - "true"
              - all
            description: "Default: false"
          required: false
          description: "Default: false"
          name: deleted
          in: query
        - schema:
            type: string
          required: false
          name: search
          in: query
      responses:
        "200":
          description: Paginated clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Clients
      operationId: createClient
      summary: Create a client (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                status:
                  type: string
                description:
                  type: string
              required:
                - name
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/clients/{id}:
    get:
      tags:
        - Clients
      operationId: getClientById
      summary: Get a client by id
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags:
        - Clients
      operationId: updateClient
      summary: Update a client (optimistic lock via updated_at)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                status:
                  type: string
                description:
                  type: string
                updated_at:
                  type: string
                  minLength: 1
              required:
                - updated_at
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Clients
      operationId: deleteClient
      summary: Soft-delete a client (admin only)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Client soft-deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/clients/{id}/restore:
    post:
      tags:
        - Clients
      operationId: restoreClient
      summary: Restore a soft-deleted client (admin only)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Client restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/people:
    get:
      tags:
        - People
      operationId: listPeople
      summary: List people
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 1
            description: "Default: 1"
          required: false
          description: "Default: 1"
          name: page
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            description: "Default: 20, max: 100"
          required: false
          description: "Default: 20, max: 100"
          name: per_page
          in: query
        - schema:
            type: string
            enum:
              - name
              - created_at
              - updated_at
            description: "Default: name"
          required: false
          description: "Default: name"
          name: sort
          in: query
        - schema:
            type: string
            enum:
              - asc
              - desc
            description: "Default: asc"
          required: false
          description: "Default: asc"
          name: order
          in: query
        - schema:
            type: string
            enum:
              - "false"
              - "true"
              - all
            description: "Default: false"
          required: false
          description: "Default: false"
          name: deleted
          in: query
        - schema:
            type: string
          required: false
          name: search
          in: query
        - schema:
            type: string
          required: false
          name: type
          in: query
      responses:
        "200":
          description: Paginated people
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - People
      operationId: createPerson
      summary: Create a person (admin or member)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                email:
                  type: string
                  format: email
                phone:
                  type: string
                type:
                  type: string
                  enum:
                    - internal_engineer
                    - vendor
                    - client_contact
                    - project_owner
                    - approver
                notes:
                  type: string
              required:
                - name
                - type
      responses:
        "201":
          description: Person created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/activity-logs:
    get:
      tags:
        - Activity
      operationId: listActivityLogs
      summary: List activity log entries (append-only audit trail; read-only — no
        write endpoints exist for this resource)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 1
            description: "Default: 1"
          required: false
          description: "Default: 1"
          name: page
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            description: "Default: 20"
          required: false
          description: "Default: 20"
          name: per_page
          in: query
        - schema:
            type: string
            enum:
              - created_at
            description: "Only sortable field. Default: created_at"
          required: false
          description: "Only sortable field. Default: created_at"
          name: sort
          in: query
        - schema:
            type: string
            enum:
              - asc
              - desc
            description: "Default: desc"
          required: false
          description: "Default: desc"
          name: order
          in: query
        - schema:
            type: string
          required: false
          name: entity_type
          in: query
        - schema:
            type: string
            format: uuid
          required: false
          name: entity_id
          in: query
        - schema:
            type: string
          required: false
          name: action
          in: query
        - schema:
            type: string
            format: uuid
            description: Filters by the people.id who made the change
          required: false
          description: Filters by the people.id who made the change
          name: changed_by
          in: query
        - schema:
            type: string
          required: false
          name: from
          in: query
        - schema:
            type: string
          required: false
          name: to
          in: query
      responses:
        "200":
          description: Paginated activity log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityLogListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"